# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: City Stories Guide API
  description: 'REST API for City Stories Guide — an ambient audio guide app that tells stories

    about nearby places while walking. Provides endpoints for discovering nearby stories,

    managing cities/POIs/stories, authentication, purchases, and admin operations.

    '
  version: 1.0.0
  contact:
    name: City Stories Guide Team
servers:
- url: /api/v1
  description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from auth endpoints
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          example: invalid request
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: abc-123-def
    ValidationDetail:
      type: object
      required:
      - field
      - message
      properties:
        field:
          type: string
          description: JSON field name that failed validation
          example: email
        message:
          type: string
          description: Human-readable validation failure reason
          example: must be a valid email address
    ValidationError:
      type: object
      required:
      - error
      - details
      properties:
        error:
          type: string
          enum:
          - validation_error
          example: validation_error
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationDetail'
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: abc-123-def
    ErrorOrValidationError:
      oneOf:
      - $ref: '#/components/schemas/Error'
      - $ref: '#/components/schemas/ValidationError'
...
    CursorPaginatedResponse:
      description: Common cursor-based pagination response fields
      type: object
      required:
      - items
      - next_cursor
      - has_more
      properties:
        items:
          type: array
          description: Page items. Concrete endpoints specify the item schema.
          items: {}
        next_cursor:
          type: string
          description: Opaque cursor to pass as cursor param for the next page. Empty string if no more pages.
          example: aWQ6NDI=
        has_more:
          type: boolean
          description: Whether there are more pages after this one
          example: true
...
    StoryResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/Story'
...
    StoriesPageResponse:
      allOf:
      - $ref: '#/components/schemas/CursorPaginatedResponse'
      - type: object
        properties:
          items:
            type: array
            items:
              $ref: '#/components/schemas/Story'
...
    Story:
      type: object
      required:
      - id
      - poi_id
      - language
      - text
      - layer_type
      - order_index
      - is_inflation
      - confidence
      - status
      - created_at
      - updated_at
      properties:
        id:
          type: integer
          example: 1
        poi_id:
          type: integer
          example: 1
        language:
          type: string
          example: en
        text:
          type: string
          example: Standing on this ancient hill, you can almost hear...
        audio_url:
          type: string
          nullable: true
          example: https://storage.example.com/stories/1/en.mp3
        duration_sec:
          type: integer
          format: int16
          nullable: true
          example: 120
        layer_type:
          type: string
          enum:
          - atmosphere
          - human_story
          - hidden_detail
          - time_shift
          - general
          example: atmosphere
        order_index:
          type: integer
          format: int16
          example: 0
        is_inflation:
          type: boolean
          example: false
        confidence:
          type: integer
          format: int16
          example: 90
        sources:
          type: object
          nullable: true
        status:
          type: string
          enum:
          - active
          - disabled
          - reported
          - pending_review
          example: active
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /stories:
    get:
      tags:
      - Stories
      summary: List stories
      operationId: listStories
      parameters:
      - name: poi_id
        in: query
        required: true
        schema:
          type: integer
      - name: language
        in: query
        schema:
          type: string
          default: en
      - name: status
        in: query
        schema:
          type: string
          enum:
          - active
          - disabled
          - reported
          - pending_review
      - name: cursor
        in: query
        description: Opaque cursor from previous response's next_cursor
        schema:
          type: string
      - name: limit
        in: query
        description: Number of items per page (default 20, max 100)
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: Paginated list of stories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoriesPageResponse'
        '400':
          description: Missing or invalid poi_id, status, cursor, or limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOrValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stories/{id}:
    get:
      tags:
      - Stories
      summary: Get story by ID
      operationId: getStory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '400':
          description: Invalid story ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Story details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryResponse'
        '404':
          description: Story not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...