# backend/internal/handler/story_handler.go
package handler

import (
	"context"
	"encoding/json"
	"errors"
	"net/http"

	"github.com/gin-gonic/gin"

	"github.com/saas/city-stories-guide/backend/internal/domain"
	"github.com/saas/city-stories-guide/backend/internal/repository"
)

// StoryRepository defines the interface for story database operations.
type StoryRepository interface {
	Create(ctx context.Context, story *domain.Story) (*domain.Story, error)
	GetByID(ctx context.Context, id int) (*domain.Story, error)
	GetByPOIID(ctx context.Context, poiID int, language string, status *domain.StoryStatus) ([]domain.Story, error)
	ListByPOIID(ctx context.Context, poiID int, language string, status *domain.StoryStatus, page domain.PageRequest, sort repository.ListSort) (*domain.PageResponse[domain.Story], error)
	Update(ctx context.Context, story *domain.Story) (*domain.Story, error)
	Delete(ctx context.Context, id int) error
}

// StoryCleanupService schedules orphaned audio cleanup on story mutations.
type StoryCleanupService interface {
	UpdateStory(ctx context.Context, story *domain.Story) (*domain.Story, error)
	DeleteStory(ctx context.Context, id int) error
	ScheduleReplacedAudio(ctx context.Context, oldAudioURL string)
}

// StoryHandler handles CRUD operations for stories.
type StoryHandler struct {
	repo    StoryRepository
	audit   AuditLogger
	cleanup StoryCleanupService
}

// NewStoryHandler creates a new StoryHandler.
func NewStoryHandler(repo StoryRepository, audit AuditLogger, cleanup ...StoryCleanupService) *StoryHandler {
	h := &StoryHandler{repo: repo, audit: audit}
	if len(cleanup) > 0 {
		h.cleanup = cleanup[0]
	}
	return h
}

// createStoryRequest represents the request body for creating a story.
type createStoryRequest struct {
	POIID       int                   `json:"poi_id" binding:"required"`
	Language    string                `json:"language" binding:"required,len=2"`
	Text        string                `json:"text" binding:"required,max=50000"`
	AudioURL    *string               `json:"audio_url" binding:"omitempty,url"`
	DurationSec *int16                `json:"duration_sec" binding:"omitempty,gte=0"`
	LayerType   domain.StoryLayerType `json:"layer_type" binding:"required,oneof=atmosphere human_story hidden_detail time_shift general"`
	OrderIndex  *int16                `json:"order_index"`
	IsInflation *bool                 `json:"is_inflation"`
	Confidence  *int16                `json:"confidence" binding:"omitempty,gte=0,lte=100"`
	Sources     *json.RawMessage      `json:"sources"`
	Status      *domain.StoryStatus   `json:"status" binding:"omitempty,oneof=active disabled reported pending_review"`
}

// updateStoryRequest represents the request body for updating a story.
type updateStoryRequest struct {
	POIID       int                   `json:"poi_id" binding:"required"`
	Language    string                `json:"language" binding:"required,len=2"`
	Text        string                `json:"text" binding:"required,max=50000"`
	AudioURL    *string               `json:"audio_url" binding:"omitempty,url"`
	DurationSec *int16                `json:"duration_sec" binding:"omitempty,gte=0"`
	LayerType   domain.StoryLayerType `json:"layer_type" binding:"required,oneof=atmosphere human_story hidden_detail time_shift general"`
	OrderIndex  *int16                `json:"order_index"`
	IsInflation *bool                 `json:"is_inflation"`
	Confidence  *int16                `json:"confidence" binding:"omitempty,gte=0,lte=100"`
	Sources     *json.RawMessage      `json:"sources"`
	Status      *domain.StoryStatus   `json:"status" binding:"omitempty,oneof=active disabled reported pending_review"`
}

var adminStorySortColumns = map[string]repository.SortColumn{
	"id":          {Key: "id", Column: "id", Type: repository.SortValueInt},
	"poi_id":      {Key: "poi_id", Column: "poi_id", Type: repository.SortValueInt},
	"language":    {Key: "language", Column: "language", Type: repository.SortValueString},
	"status":      {Key: "status", Column: "status", Type: repository.SortValueString},
	"layer_type":  {Key: "layer_type", Column: "layer_type", Type: repository.SortValueString},
	"order_index": {Key: "order_index", Column: "order_index", Type: repository.SortValueInt16},
	"confidence":  {Key: "confidence", Column: "confidence", Type: repository.SortValueInt16},
	"created_at":  {Key: "created_at", Column: "created_at", Type: repository.SortValueTime},
	"updated_at":  {Key: "updated_at", Column: "updated_at", Type: repository.SortValueTime},
}

// ListStories handles GET /api/v1/stories?poi_id=&language=&status=.
func (h *StoryHandler) ListStories(c *gin.Context) {
	h.listStories(c, repository.ListSort{By: "id", Dir: repository.SortDirAsc})
}

// ListAdminStories handles GET /api/v1/admin/stories?poi_id=&language=&status=.
func (h *StoryHandler) ListAdminStories(c *gin.Context) {
	sortReq, ok := parseListSort(c, adminStorySortColumns, "id", repository.SortDirAsc)
	if !ok {
		return
	}

	h.listStories(c, sortReq)
}

func (h *StoryHandler) listStories(c *gin.Context, sortReq repository.ListSort) {
	poiID, ok := parseRequiredQueryInt(c, "poi_id")
	if !ok {
		return
	}

	language := c.DefaultQuery("language", "en")

	var statusFilter *domain.StoryStatus
	if s := c.Query("status"); s != "" {
		if err := domain.ValidateEnum(s, "status", []string{
			string(domain.StoryStatusActive),
			string(domain.StoryStatusDisabled),
			string(domain.StoryStatusReported),
			string(domain.StoryStatusPendingReview),
		}); err != nil {
			validationErrorResponse(c, err)
			return
		}
		st := domain.StoryStatus(s)
		statusFilter = &st
	}

	pageReq, ok := parseCursorPagination(c)
	if !ok {
		return
	}

	result, err := h.repo.ListByPOIID(c.Request.Context(), poiID, language, statusFilter, pageReq, sortReq)
	if err != nil {
		if isCursorError(err) {
			errorJSON(c, http.StatusBadRequest, err.Error())
			return
		}
		errorJSON(c, http.StatusInternalServerError, "failed to fetch stories")
		return
	}

	writeCursorPage(c, result)
}

// GetStory handles GET /api/v1/stories/:id.
func (h *StoryHandler) GetStory(c *gin.Context) {
	id, ok := parseIDParam(c)
	if !ok {
		return
	}

	story, err := h.repo.GetByID(c.Request.Context(), id)
	if err != nil {
		if errors.Is(err, repository.ErrNotFound) {
			errorJSON(c, http.StatusNotFound, "story not found")
			return
		}
		errorJSON(c, http.StatusInternalServerError, "failed to fetch story")
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": story})
}

// CreateStory handles POST /api/v1/admin/stories.
func (h *StoryHandler) CreateStory(c *gin.Context) {
	var req createStoryRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		validationErrorResponse(c, err)
		return
	}
	if err := validateStoryRequest(req.Language); err != nil {
		validationErrorResponse(c, err)
		return
	}

	orderIndex := int16(0)
	if req.OrderIndex != nil {
		orderIndex = *req.OrderIndex
	}

	isInflation := false
	if req.IsInflation != nil {
		isInflation = *req.IsInflation
	}

	confidence := int16(80)
	if req.Confidence != nil {
		confidence = *req.Confidence
	}

	status := domain.StoryStatusActive
	if req.Status != nil {
		status = *req.Status
	}

	var sources json.RawMessage
	if req.Sources != nil {
		sources = *req.Sources
	}

	story := &domain.Story{
		POIID:       req.POIID,
		Language:    req.Language,
		Text:        req.Text,
		AudioURL:    req.AudioURL,
		DurationSec: req.DurationSec,
		LayerType:   req.LayerType,
		OrderIndex:  orderIndex,
		IsInflation: isInflation,
		Confidence:  confidence,
		Sources:     sources,
		Status:      status,
	}

	created, err := h.repo.Create(c.Request.Context(), story)
	if err != nil {
		if handleDBError(c, repository.ClassifyError(err), "story") {
			return
		}
		errorJSON(c, http.StatusInternalServerError, "failed to create story")
		return
	}

	auditEntry(c, h.audit, "create", "story", resourceID(created.ID), storyAuditPayload(req.POIID, req.Language, string(req.LayerType)))
	c.JSON(http.StatusCreated, gin.H{"data": created})
}

// UpdateStory handles PUT /api/v1/admin/stories/:id.
func (h *StoryHandler) UpdateStory(c *gin.Context) {
	id, ok := parseIDParam(c)
	if !ok {
		return
	}

	var req updateStoryRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		validationErrorResponse(c, err)
		return
	}
	if err := validateStoryRequest(req.Language); err != nil {
		validationErrorResponse(c, err)
		return
	}

	orderIndex := int16(0)
	if req.OrderIndex != nil {
		orderIndex = *req.OrderIndex
	}

	isInflation := false
	if req.IsInflation != nil {
		isInflation = *req.IsInflation
	}

	confidence := int16(80)
	if req.Confidence != nil {
		confidence = *req.Confidence
	}

	status := domain.StoryStatusActive
	if req.Status != nil {
		status = *req.Status
	}

	var sources json.RawMessage
	if req.Sources != nil {
		sources = *req.Sources
	}

	story := &domain.Story{
		ID:          id,
		POIID:       req.POIID,
		Language:    req.Language,
		Text:        req.Text,
		AudioURL:    req.AudioURL,
		DurationSec: req.DurationSec,
		LayerType:   req.LayerType,
		OrderIndex:  orderIndex,
		IsInflation: isInflation,
		Confidence:  confidence,
		Sources:     sources,
		Status:      status,
	}

	var (
		updated *domain.Story
		err     error
	)
	if h.cleanup != nil {
		updated, err = h.cleanup.UpdateStory(c.Request.Context(), story)
	} else {
		updated, err = h.repo.Update(c.Request.Context(), story)
	}
	if err != nil {
		if handleDBError(c, repository.ClassifyError(err), "story") {
			return
		}
		errorJSON(c, http.StatusInternalServerError, "failed to update story")
		return
	}

	auditEntry(c, h.audit, "update", "story", resourceID(id), storyAuditPayload(req.POIID, req.Language, string(req.LayerType)))
	c.JSON(http.StatusOK, gin.H{"data": updated})
}

// DeleteStory handles DELETE /api/v1/admin/stories/:id.
func (h *StoryHandler) DeleteStory(c *gin.Context) {
	id, ok := parseIDParam(c)
	if !ok {
		return
	}

	var err error
	if h.cleanup != nil {
		err = h.cleanup.DeleteStory(c.Request.Context(), id)
	} else {
		err = h.repo.Delete(c.Request.Context(), id)
	}
	if err != nil {
		if handleDBError(c, repository.ClassifyError(err), "story") {
			return
		}
		errorJSON(c, http.StatusInternalServerError, "failed to delete story")
		return
	}

	auditEntry(c, h.audit, "delete", "story", resourceID(id), nil)
	c.JSON(http.StatusOK, gin.H{"message": "story deleted"})
}

func validateStoryRequest(language string) error {
	return domain.ValidateISO639_1(language, "language")
}